requirements.txt
python-dateutil==2.8.1
requests==2.25.1
schedule==1.1.0
paho-mqtt==1.5.1
